feat: Move index to components folder
