+++ src/components/__test__/index.test.tsx
@@ -5,7 +5,7 @@ import { useTranslationContext } from 'hooks'
 import { Locales } from 'types'
 import { setBrowserLanguage } from 'utils'
 
-import TranslationContainer from '..'
+import TranslationContainer from '../TranslationContainer'
 
 const locales = {
   'pt-BR': {
